requirement.txt
flask
pymongo
python-dotenv
flask-jwt-extended